split up providers
