+++ packages/ui/src/common/checkbox-field.tsx
@@ -1,7 +1,7 @@
 import * as Common from '@clerk/elements/common';
 import React from 'react';
 
-import * as Field from '../primitives/field';
+import * as Field from '~/primitives/field';
 
 export function CheckboxField({
   label,
