+++ packages/ui/src/constants/localizations.ts
@@ -1,0 +1,5 @@
+// As part of https://linear.app/clerk/issue/SDKI-191
+export const LOCALIZATION_NEEDED = {
+  formFieldAccessibleLabel__emailOrUsernameOrPhone: 'Toggle between email or username, and phone.',
+  formFieldAccessibleLabel__phoneOrUsername: 'Toggle between phone and username.',
+};
